+++ src/modules/coordinate.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, Copy)]
 pub struct Point {
-    pub col: u16,
-    pub row: u16,
+    pub col: usize,
+    pub row: usize,
 }
